Move inline extension logic in to browser
