+++ browser/src/shared/platform/context.ts
@@ -18,6 +18,9 @@ import { editClientSettings, fetchViewerSettings, mergeCascades, storageSettings
 import { requestGraphQlHelper } from '../backend/requestGraphQl'
 import { isHTTPAuthError } from '../../../../shared/src/backend/fetch'
 import { asError } from '../../../../shared/src/util/errors'
+import { InlineExtensionsService, shouldUseInlineExtensions } from './inlineExtensionsService'
+import { ModelService } from '../../../../shared/src/api/client/services/modelService'
+import { IExtensionsService, ExtensionsService } from '../../../../shared/src/api/client/services/extensionsService'
 
 export interface SourcegraphIntegrationURLs {
     /**
@@ -167,6 +170,15 @@ export function createPlatformContext(
         sideloadedExtensionURL: isInPage
             ? new LocalStorageSubject<string | null>('sideloadedExtensionURL', null)
             : new ExtensionStorageSubject('sideloadedExtensionURL', null),
+        createExtensionsService(modelService: Pick<ModelService, 'activeLanguages'>): IExtensionsService {
+            // On Firefox extension, use the inline extensions service.
+            if (shouldUseInlineExtensions()) {
+                return new InlineExtensionsService()
+            }
+
+            // On all other platforms, use the standard extensions service.
+            return new ExtensionsService(this, modelService)
+        },
     }
     return context
 }
