+++ shared/src/api/client/services.ts
@@ -6,7 +6,7 @@ import { CompletionItemProviderRegistry } from './services/completion'
 import { ContributionRegistry } from './services/contribution'
 import { TextDocumentDecorationProviderRegistry } from './services/decoration'
 import { createViewerService } from './services/viewerService'
-import { createExtensionsService } from './services/extensionsService'
+import { IExtensionsService, ExtensionsService } from './services/extensionsService'
 import { TextDocumentHoverProviderRegistry } from './services/hover'
 import { LinkPreviewProviderRegistry } from './services/linkPreview'
 import { TextDocumentLocationProviderIDRegistry, TextDocumentLocationProviderRegistry } from './services/location'
@@ -29,8 +29,15 @@ export class Services {
             | 'getScriptURLForExtension'
             | 'clientApplication'
             | 'sideloadedExtensionURL'
+            | 'createExtensionsService'
         >
-    ) {}
+    ) {
+        if (platformContext.createExtensionsService) {
+            this.extensions = platformContext.createExtensionsService(this.model)
+        } else {
+            this.extensions = new ExtensionsService(platformContext, this.model)
+        }
+    }
 
     public readonly commands = new CommandRegistry()
     public readonly context = createContextService(this.platformContext)
@@ -44,7 +51,7 @@ export class Services {
         this.platformContext.settings,
         this.context.data
     )
-    public readonly extensions = createExtensionsService(this.platformContext, this.model)
+    public readonly extensions: IExtensionsService
     public readonly linkPreviews = new LinkPreviewProviderRegistry()
     public readonly textDocumentDefinition = new TextDocumentLocationProviderRegistry()
     public readonly textDocumentReferences = new TextDocumentLocationProviderRegistry<ReferenceParams>()
